+plasma-sdk

+++ create-kde.sh
@@ -7,12 +7,15 @@ while read -r line; do
 
     pkg=$(       echo $file|sed 's|^.*/||')    # Remove directory
     name=$(      echo $pkg |sed 's|-6.*$||')   # Isolate package name
+    if ! [[ -f /var/lib/book-packages/$name ]]; then
+	    continue
+    fi
     source add_deps.sh
-    cp -r attica $name
-    sed -i -e "s|attica|$name|g" $name/build.sh
+    cp -r kdecoration $name
+    sed -i -e "s|kdecoration|$name|g" $name/build.sh
     mv /var/lib/book-packages/$name /var/lib/custom-packages/$name
     add_deps $name
 
-done < frameworks-6.29.0.md5
+done < plasma-6.7.4.md5
 
 exit
